Termine los tests con el maximo de coverage que pude, modifique la manera de trabajar las cartas en el entrenador dado que entendi el enunciado mal previamente. Aplique double dispatch para poder asi lograr ver como jugar las cartas. Faltaria hacer las tareas burocraticas y era

+++ src/cc3002.tarea1/PokemonTypes/FighterPokemon.java
@@ -3,7 +3,7 @@ import cc3002.tarea1.Pokemon;
 import cc3002.tarea1.ISkill;
 import java.util.ArrayList;
 public class FighterPokemon extends Pokemon{
-    public FighterPokemon(String name, int id, double healthPoints, ArrayList<ISkill> skills){
+    public FighterPokemon(String name, int id, int healthPoints, ArrayList<ISkill> skills){
         super(name, id, healthPoints, skills);
     }
     public void attack(Pokemon A){
